+++ src/applications/health-care-questionnaire/questionnaire/tests/unit/utils/createFakeStores.js
@@ -221,43 +221,6 @@ const createFakeReasonForVisitStore = ({ reason = '' }) => {
   };
 };
 
-const createFakeConfirmationStore = ({ hasData }) => {
-  return {
-    getState: () => {
-      return hasData
-        ? {
-            form: {
-              submission: {
-                response: {
-                  veteranInfo: { fullName: 'Mickey Mouse' },
-                  timestamp: new Date(),
-                },
-              },
-            },
-            questionnaireData: {
-              context: {
-                appointment: {
-                  attributes: {
-                    vdsAppointments: [
-                      {
-                        clinic: {
-                          stopCode: '323',
-                          facility: { displayName: 'Magic Kingdom' },
-                        },
-                      },
-                    ],
-                  },
-                },
-              },
-            },
-          }
-        : { form: {}, questionnaireData: {} };
-    },
-    subscribe: () => {},
-    dispatch: () => {},
-  };
-};
-
 const createFakeIntroductionPageStore = (formId, savedForms) => {
   return {
     getState: () => ({
@@ -313,7 +276,6 @@ export {
   createFakeUserStore,
   createFakeReasonForVisitDescriptionStore,
   createFakeReasonForVisitStore,
-  createFakeConfirmationStore,
   createFakeIntroductionPageStore,
   createFakeStopCodeStore,
   createFakeExpiresAtStore,
